+++ src/main/kotlin/homeworks/first/utils/vo/SolutionInfo.kt
@@ -1,9 +1,11 @@
 package homeworks.first.utils.vo
 
+import java.math.BigDecimal
+
 data class SolutionInfo(
     val methodName: String,
-    val result: Double,
-    val seqLength: Double,
-    val absoluteDiscrepancy: Double,
+    val result: BigDecimal,
+    val seqLength: BigDecimal,
+    val absoluteDiscrepancy: BigDecimal,
     val stepsNumber: Int
 )
