+++ src/main/kotlin/homeworks/second/methods/InterpolationMethod.kt
@@ -1,17 +1,18 @@
 package homeworks.second.methods
 
+import java.math.BigDecimal
 import java.util.SortedMap
 
 interface InterpolationMethod {
-    val table: Map<Double, Double>
+    val table: Map<BigDecimal, BigDecimal>
 
     val degree: Int
 
-    fun setUp(table: SortedMap<Double, Double>, degree: Int)
+    fun setUp(table: SortedMap<BigDecimal, BigDecimal>, degree: Int)
 
     fun evaluate(
-        point: Double
-    ): Double
+        point: BigDecimal
+    ): BigDecimal
 
     fun getMethodName(): String
 }
