+++ src/main/kotlin/homeworks/utils/vo/Seq.kt
@@ -1,8 +1,10 @@
 package homeworks.utils.vo
 
+import java.math.BigDecimal
+
 data class Seq(
-    val left: Double,
-    val right: Double
+    val left: BigDecimal,
+    val right: BigDecimal
 ) {
     init {
         require(right >= left)
